Show how to use the MemoryFSHandler and a custom WebViewHandler
