+++ samples/html2/webview_sample.html
@@ -1,0 +1,29 @@
+<html>
+<head>
+<title>Home</title>
+</head>        
+<body>
+<h1>html2.WebView Sample</h1>
+<p>This is a super simple launch page for the html2.WebView sample.</p>
+
+<br>
+<p>Normal web pages and real web sites work just as expected:
+<ul>
+        <li><a href="https://wxPython.org/">wxPython Home</a></li>
+        <li><a href="https://wxWidgets.org/">wxWidgets Home</a></li>
+        <li><a href="https://www.microsoft.com/">Microsoft Home</a></li>
+        <li><a href="http://google.com">Google Home</a></li>
+</ul>
+</p>
+
+<br>
+<p>
+Resources can also be loaded from custom file systems:
+<ul>
+        <li><a href="memory:page1.htm">wx.MemoryFSHandler</a></li>
+        <li>Custom <a href="custom:page1.html">wx.html2.WebViewHandler</a></li>
+</ul>
+</p>
+<p></p>
+</body>
+</html>
